+++ CHANGELOG.md
@@ -5,6 +5,8 @@ See updating [Changelog example here](https://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 
+## [2.11.0] - 2023-06-07
+
 ### Added
 - kubernetes: `private_node_groups` field to `upcloud_kubernetes_cluster` resource
 - server: properties `timezone`, `video_model` and `nic_model`
@@ -362,7 +364,8 @@ Updated upcloud-go-api, added build/CI scripts, and repackaged 0.1.0 as 1.0.0.
 - resource_upcloud_firewall_rule removed and replaced by resource_upcloud_firewall_rules
 - resource_upcloud_zone removed and replaced by zone and zones datasources
 
-[Unreleased]: https://github.com/UpCloudLtd/terraform-provider-upcloud/compare/v2.10.0...HEAD
+[Unreleased]: https://github.com/UpCloudLtd/terraform-provider-upcloud/compare/v2.11.0...HEAD
+[2.11.0]: https://github.com/UpCloudLtd/terraform-provider-upcloud/compare/v2.10.0...v2.11.0
 [2.10.0]: https://github.com/UpCloudLtd/terraform-provider-upcloud/compare/v2.9.1...v2.10.0
 [2.9.1]: https://github.com/UpCloudLtd/terraform-provider-upcloud/compare/v2.9.0...v2.9.1
 [2.9.0]: https://github.com/UpCloudLtd/terraform-provider-upcloud/compare/v2.8.4...v2.9.0